feat: delete /medias/id

+++ src/medias/medias.service.ts
@@ -1,15 +1,20 @@
 import {
   ConflictException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
 import { CreateMediaDto } from './dto/create-media.dto';
 import { UpdateMediaDto } from './dto/update-media.dto';
 import { MediasRepository } from './medias.repository';
+import { PublicationRepository } from 'src/publication/publication.repository';
 
 @Injectable()
 export class MediasService {
-  constructor(private readonly repository: MediasRepository) {}
+  constructor(
+      private readonly repository: MediasRepository,
+      private readonly repositoryPublication: PublicationRepository
+    ) {}
 
   async create(createMediaDto: CreateMediaDto) {
     const { title, username } = createMediaDto;
@@ -57,9 +62,14 @@ export class MediasService {
 
   async remove(id: number) {
     const media = await this.repository.findOne(id);
+    const publication = await this.repositoryPublication.publicationByMedeaId(media.id);
 
     if (!media) {
       throw new NotFoundException('Midia não encontrada!');
+    }
+
+    if(publication > 0){
+      throw new ForbiddenException('Esta mídia está vinculada a uma publicação!')
     }
     
     return await this.repository.remove(id);
